module: Configurable fault logging
Introduces the /proc/krf/log_faults file, which can
be read from and written to to control logging
of faulty syscalls to the kernel message buffer.

By default, no logging is performed.

This could probably have been done via a bit
set in the process personality(2) instead, but this
is a reasonable enough approach for now.