README: Use installed paths instead of src paths (#71)
Since we recommend doing `sudo make install` on Linux or `sudo gmake install-utils` on FreeBSD (I assume it does indeed put tools to `/usr/local/bin` and that this path is usually in `$PATH` on that system) there is no reason to show examples with dev-like paths e.g. `./src/krfexec/krfexec`.